Detect encoding problems (#108)
SDCcc currently accepts the encoding specified in a message and uses it
to decode the message. When the encoding is specified incorrectly or the
message contains characters that are not mappable using the specified
encoding, then they are replaced by Charset-specific default characters,
which may lead to False-positives or False-negatives in Test Cases that
rely on messages stored in the database.

A much better approach would be to check the encoding while decoding the
messages and to raise an Encoding Error when decoding a message with the
specified encoding fails. As usual, these Encoding Errors can be managed
using the EnableMessageEncodingCheck and SummarizeMessageEncodingErrors
configuration options.

# Checklist

The following aspects have been respected by the author of this pull
request, confirmed by both pull request assignee **and** reviewer:

* Adherence to coding conventions
  * [x] Pull Request Assignee
  * [x] Reviewer
* Adherence to javadoc conventions
  * [x] Pull Request Assignee
  * [x] Reviewer
* Changelog update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* README update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* config update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* SDCcc executable ran against a test device (if necessary)
  * [x] Pull Request Assignee
  * [x] Reviewer

---------

Co-authored-by: Maximilian Pilz <maximilian.pilz.git@gmail.com>