Let DescriptionModificationUptPrecondition use the new TriggerAnyDescriptorUpdate manipulation (#117)
The DescriptionModificationUptPrecondition previously had the problem
that it did not work with Devices that did not support Updating MDS
descriptors.
In order to remove this assumption, I adapted it to use the new
TriggerAnyDescriptorUpdate manipulation, which allows the device to
choose which descriptor it updates. It should hence succeed as long as
the device has some descriptor that can be updated and the Precondition
should hence now work with all devices whose T2IAPI-Server implements
this new Manipulation.

NOTE: this code only compiles once the
[PR#50](https://github.com/Draegerwerk/t2iapi/pull/50) in T2IAPI has
been merged to master. I might have to adjust the version of T2IAPI in
the pom.xml to use the correct version after that happened...

# Checklist

The following aspects have been respected by the author of this pull
request, confirmed by both pull request assignee **and** reviewer:

* Adherence to coding conventions
  * [x] Pull Request Assignee
  * [x] Reviewer
* Adherence to javadoc conventions
  * [x] Pull Request Assignee
  * [x] Reviewer
* Changelog update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* README update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* config update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* SDCcc executable ran against a test device (if necessary)
  * [x] Pull Request Assignee
  * [x] Reviewer