fix: changed biceps:R5003 test case (#232)
In the mdib representation on the consumer side in sdc-ri, when
processing description modifications, the values
for the MdDescriptionVersion and MdStateVersion are set to the implied
value if the description modification
does not contain any values for them. This is the case if the versions
are missing in the GetMdibResponse,
or generally for DescriptionModificationReports. If the
MdDescriptionVersion or MdStateVersion in the
GetMdibResponse is set to a value other than the implied value, they
would be set to the implied value again in the next
DescriptionModificationReport, which would cause the test to fail. As
the actual values are not communicated via
reports, the check of the MdDescriptionVersion and MdStateVersion has
been removed from the test implementation.

# Checklist

The following aspects have been respected by the author of this pull
request, confirmed by both pull request assignee **and** reviewer:

* Adherence to coding conventions
  * [x] Pull Request Assignee
  * [x] Reviewer
* Adherence to javadoc conventions
  * [x] Pull Request Assignee
  * [x] Reviewer
* Changelog update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* README update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* config update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* SDCcc executable ran against a test device (if necessary)
  * [x] Pull Request Assignee
  * [x] Reviewer